requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blinker==1.9.0
certifi==2025.4.26
chardet==3.0.4
click==8.2.0
compressed-rtf==1.0.7
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
extract-msg==0.29.0
flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imapclient==2.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
lxml==5.4.0
markupsafe==3.0.2
numpy==2.2.5
olefile==0.47
openai==1.78.1
packaging==25.0
pdfminer-six==20191110
pillow==11.2.1
pluggy==1.5.0
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-core==2.33.2
pymupdf==1.25.5
pytesseract==0.3.13
pytest==8.3.3
pytest-mock==3.14.0
python-docx==1.1.2
python-pptx==0.6.23
ruff==0.11.9
scikit-learn==1.6.1
scipy==1.15.3
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
speechrecognition==3.8.1
textract==1.6.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzlocal==5.3.1
werkzeug==3.1.3
xlrd==1.2.0
xlsxwriter==3.2.3
